Complete Weight Rig runtime split
